+++ arch/i386/mod.rs
@@ -5,7 +5,7 @@ pub mod debug;
 pub mod paging;
 pub mod multiboot;
 
-pub static PAGE_SIZE: uint = 0x1000;
+pub static PAGE_SIZE: usize = 0x1000;
 
 pub fn init() {
     multiboot::init();
